refactor(fusain): align naming with CDDL spec and add net_buf API
Naming consistency (CDDL alignment):
- Rename fusain_data_temperature_t fields:
  - temp → reading
  - ctrl_rpm_by_temp → temperature_rpm_control
  - target_temp → target_temperature
- Rename fusain_create_set_mode() → fusain_create_state_command()
  - Parameter: uint32_t parameter → int32_t argument

New features:
- Add net_buf decoder API for Zephyr (fusain_decode_byte_to_net_buf)
- Add fusain_packet_from_buf() helper
- Add data helper functions:
  - fusain_create_motor_data()
  - fusain_create_pump_data()
  - fusain_create_glow_data()
  - fusain_create_temp_data()
  - fusain_create_error_invalid_cmd()
  - fusain_create_error_state_reject()

Testing improvements:
- Add standalone fuzz test support (FUSAIN_FUZZ_ENABLED)
- Add standalone-fuzz task with configurable rounds/seed
- Include standalone fuzz in CI (1M rounds)
- Enable -Werror for test builds
- Add net_buf API tests (Zephyr only)
- Maintain 100% code coverage on fusain.c

Build system:
- Add CONFIG_FUSAIN_NET_BUF Kconfig option (default y)
- Add coverage-check task for Zephyr tests
- Remove ignore_error workaround from coverage-check

Wire compatibility: Unchanged (CBOR uses integer keys)

+++ tests/src/test_crc.c
@@ -46,17 +46,17 @@ ZTEST(fusain_crc, test_crc_various_lengths)
   uint8_t data[FUSAIN_MAX_PAYLOAD_SIZE];
 
   /* Fill with pattern */
-  for (int i = 0; i < sizeof(data); i++) {
-    data[i] = i & 0xFF;
+  for (size_t i = 0; i < sizeof(data); i++) {
+    data[i] = (uint8_t)(i & 0xFF);
   }
 
   /* Test different lengths */
   uint16_t prev_crc = 0xFFFF;
-  for (int len = 1; len <= sizeof(data); len++) {
+  for (size_t len = 1; len <= sizeof(data); len++) {
     uint16_t crc = fusain_crc16(data, len);
     /* Each length should produce a different CRC */
     zassert_not_equal(crc, prev_crc,
-        "Different lengths should produce different CRCs (len=%d)",
+        "Different lengths should produce different CRCs (len=%zu)",
         len);
     prev_crc = crc;
   }
